requirements.txt
Authlib==1.6.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
et_xmlfile==2.0.0
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.3.1
openpyxl==3.1.5
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
